remove depre code. add code to convert old models.
