+++ experiments/scripts/convert_vgg16.sh
@@ -1,0 +1,70 @@
+#!/bin/bash
+
+set -x
+set -e
+
+export PYTHONUNBUFFERED="True"
+
+GPU_ID=$1
+DATASET=$2
+NET=vgg16
+
+array=( $@ )
+len=${#array[@]}
+EXTRA_ARGS=${array[@]:2:$len}
+EXTRA_ARGS_SLUG=${EXTRA_ARGS// /_}
+
+case ${DATASET} in
+  pascal_voc)
+    TRAIN_IMDB="voc_2007_trainval"
+    TEST_IMDB="voc_2007_test"
+    STEPSIZE=50000
+    ITERS=70000
+    ANCHORS="[8,16,32]"
+    RATIOS="[0.5,1,2]"
+    ;;
+  pascal_voc_0712)
+    TRAIN_IMDB="voc_2007_trainval+voc_2012_trainval"
+    TEST_IMDB="voc_2007_test"
+    STEPSIZE=80000
+    ITERS=110000
+    ANCHORS="[8,16,32]"
+    RATIOS="[0.5,1,2]"
+    ;;
+  coco)
+    TRAIN_IMDB="coco_2014_train+coco_2014_valminusminival"
+    TEST_IMDB="coco_2014_minival"
+    STEPSIZE=350000
+    ITERS=490000
+    ANCHORS="[4,8,16,32]"
+    RATIOS="[0.5,1,2]"
+    ;;
+  *)
+    echo "No dataset given"
+    exit
+    ;;
+esac
+
+set +x
+NET_FINAL=${NET}_faster_rcnn_iter_${ITERS}
+set -x
+
+if [ ! -f ${NET_FINAL}.index ]; then
+    if [[ ! -z  ${EXTRA_ARGS_SLUG}  ]]; then
+        CUDA_VISIBLE_DEVICES=${GPU_ID} time python ./tools/convert_from_depre.py \
+            --snapshot ${NET_FINAL} \
+            --imdb ${TRAIN_IMDB} \
+            --iters ${ITERS} \
+            --cfg experiments/cfgs/${NET}.yml \
+            --tag ${EXTRA_ARGS_SLUG} \
+            --set ANCHOR_SCALES ${ANCHORS} ANCHOR_RATIOS ${RATIOS} TRAIN.STEPSIZE ${STEPSIZE} ${EXTRA_ARGS}
+    else
+        CUDA_VISIBLE_DEVICES=${GPU_ID} time python ./tools/convert_from_depre.py \
+            --snapshot ${NET_FINAL} \
+            --imdb ${TRAIN_IMDB} \
+            --iters ${ITERS} \
+            --cfg experiments/cfgs/${NET}.yml \
+            --set ANCHOR_SCALES ${ANCHORS} ANCHOR_RATIOS ${RATIOS} TRAIN.STEPSIZE ${STEPSIZE} ${EXTRA_ARGS}
+    fi
+fi
+
